requirements.txt
appium-python-client
selenium
pillow